Prioritize icons vs check marks in menu items
This commit changes Phantom menu item philosophy:

  Phantom treats every menu item with the same space
  on the left for a check mark, even if it doesn't
  have the checkable property.

To this instead:

  Phantom treats every menu item with the same space
  on the left for an icon, even if it doesn't
  have an icon.

Previously the icons would have been merged with the text and it would
look very weird in my eyes.

The check marks will be rendred in the space of the icon.

The only weird case remains when the menu item has both an icon and a
check mark.