Modernize the scrollbar
The scrollbar slide thumb is now a thick line with rounded ends.

The slide thumb grows wider when  the scrollbar is hovered.

The color of the slide thumb is the the color of the arrows (indicator)
for the scrollbar.

The indicator color is visible in a dark theme environment.

This mimics the behavior that I have seen with the Windows Vista style.