Improved handling of menu items with icon and check mark
Previously both would have been rendered, next to each other.

This commit will mark the icon as disabled when the check is off, and
normal icon when the check is on.

The check is now also rendered centered in the space that the icon
takes. Since now the icons are considered always there, and not the
check marks.