Modernized the TreeView expand / close arrows
I've used a triangle outline and not a full triangle. This is what macOS
and Windows are doing.

Enabled the full line selectiong for the TreeView items. Similarly to
macOS and Fusion style.

The height of the TreeView items has a "+2" like macOS and Windows
styles. This makes the TreeView not seem cramped.

And finally removed the custom TreeView identation and used the
QCommonStyle value. This also contributed to the spacious view of the
TreeView.

With these changes in place Phantom is behaving like modern macOS or
Windows.